requirements.txt
anthropic==0.96.0
fastapi==0.136.0
nltk==3.9.4
numpy==2.4.4
openai==2.32.0
praw==7.8.1
pydantic_settings==2.14.0
python-dotenv==1.2.2
Requests==2.33.1
sentence_transformers==5.4.1
uvicorn==0.44.0
